move camera into own file
