+++ src/rendering/mod.rs
@@ -1,4 +1,6 @@
-use crate::math::{UVec2, Vec2};
+use crate::math::UVec2;
+
+mod camera;
 
 struct Renderer<'a> {
     surface: wgpu::Surface<'a>,
@@ -15,23 +17,4 @@ struct Window<'a> {
     sdl_window: &'a sdl3::video::Window,
     sdl_context: sdl3::Sdl,
     size: UVec2,
-}
-
-struct Camera {
-    window_size: UVec2,
-    position: Vec2,
-    zoom: f32,
-}
-
-#[repr(C)]
-#[derive(Clone, Copy, bytemuck::Zeroable, bytemuck::Pod)]
-struct CameraUniform {
-    window_size: [u32; 2],
-    position: [f32; 2],
-    zoom: f32,
-}
-
-struct CameraBuffer {
-    buffer: wgpu::Buffer,
-    bind_group: wgpu::BindGroup,
 }
